Extend error message if "unable to detect version"
In some cases it's not possible to include the .git subdirectory and the version number doesn't matter. Then, the error message "unable to detect version, make sure your git repository is intact" doesn't help. Add a note on SETUPTOOLS_SCM_PRETEND_VERSION so users spend less time googling for a workaround.

(I do understand that this is not the recommended use-case, but there are situations in which it cannot be avoided without even uglier workarounds.)